requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
Pillow==10.0.1
gunicorn==21.2.0
EOL
